+++ packages/stream_chat_flutter/lib/src/scroll_view/draft_scroll_view/stream_draft_list_tile.dart
@@ -136,6 +136,7 @@ class DraftMessageContent extends StatelessWidget {
         StreamTimestamp(
           date: draft.createdAt.toLocal(),
           style: theme.draftTimestampStyle,
+          formatter: theme.draftTimestampFormatter,
         ),
       ],
     );